Add files via upload
My goal is to build a model that can predict the mpg (miles per gallon) based on the numerical features ("origin," "cylinders," "displacement," "horsepower," and "weight"). This can be done through a process known as regression. By training a machine learning model on historical data where the mpg values are known, the model can learn patterns and relationships between the numerical features and the mpg. Once trained, the model can then be used to predict the mpg for new instances or cars based on their numerical attributes. I applied the Random Forest Regressor algorithm to develop a machine-learning model as part of my project. To optimize the performance of the Random Forest Regressor, I carefully tuned its hyperparameters.